Update automotive demo to match 1.0 HW mappings
This includs updating the addresses of GPIO ranges used, changing the
input GPIO mappings for the joystick to match the new GPIO input
mappings after pinmux, now using PMOD0 pin 1 instead of the Mikrobus
Interrupt as a GPIO digital input for the digital pedal demo (as that is
no longer mapped to be anything), and changing the legacy ethernet and
LCD driver code to now use SPI Chip selects instead of GPIO pins.

+++ examples/automotive/cheri/receive.cc
@@ -3,7 +3,6 @@
 
 #include <compartment.h>
 #include <debug.hh>
-#include <platform-gpio.hh>
 #include <platform-ethernet.hh>
 #include <platform-gpio.hh>
 #include <platform-pwm.hh>
@@ -88,8 +87,8 @@ uint64_t wait_with_input(const uint64_t EndTime, bool *flagReset)
 	while (currentTime < EndTime)
 	{
 		currentTime = rdcycle64();
-		const uint8_t JoystickState =
-		  static_cast<uint8_t>(gpio->read_joystick());
+		const uint16_t JoystickState =
+		  static_cast<uint16_t>(gpio->read_joystick());
 		*flagReset |= joystick_in_direction(JoystickState, Pressed);
 	}
 	return currentTime;
